Improve default flyspell behavior by defining a custom predicate
salt-mode is a text-mode, not a prog-mode; most global-on flyspell
configurations will enable full spellchecking, which is not very
useful. However, even flyspell-prog-mode is not a good default, since
strings in SLS files are usually just some other configuration file
format, or a minion target, etc., which should not be spellchecked.
Instead, we want to restrict spellchecking by default to comments
and documentation only.